refactor: BoardDetailscreen
